+++ exporters/etw/test/etw_logger_test.cc
@@ -48,7 +48,8 @@ TEST(ETWLogger, LoggerCheckWithBody)
   std::string providerName = kGlobalProviderName;  // supply unique instrumentation name here
   exporter::etw::LoggerProvider lp;
 
-  auto logger        = lp.GetLogger(providerName);
+  const std::string schema_url{"https://opentelemetry.io/schemas/1.2.0"};
+  auto logger        = lp.GetLogger(providerName, "", schema_url);
   Properties attribs = {{"attrib1", 1}, {"attrib2", 2}};
   EXPECT_NO_THROW(
       logger->Log(opentelemetry::logs::Severity::kDebug, "My Log", "This is test log body"));
@@ -90,7 +91,8 @@ TEST(ETWLogger, LoggerCheckWithAttributes)
   std::string providerName = kGlobalProviderName;  // supply unique instrumentation name here
   exporter::etw::LoggerProvider lp;
 
-  auto logger = lp.GetLogger(providerName);
+  const std::string schema_url{"https://opentelemetry.io/schemas/1.2.0"};
+  auto logger = lp.GetLogger(providerName, "", schema_url);
   // Log attributes
   Properties attribs = {{"attrib1", 1}, {"attrib2", 2}};
   EXPECT_NO_THROW(logger->Log(opentelemetry::logs::Severity::kDebug, "My Log", attribs));