+++ exporters/zipkin/include/opentelemetry/exporters/zipkin/zipkin_exporter.h
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include "opentelemetry/common/spin_lock_mutex.h"
 #include "opentelemetry/ext/http/client/http_client_factory.h"
 #include "opentelemetry/ext/http/common/url_parser.h"
 #include "opentelemetry/sdk/common/env_variables.h"
@@ -82,21 +83,30 @@ public:
    * @param timeout an optional timeout, default to max.
    */
   bool Shutdown(
-      std::chrono::microseconds timeout = std::chrono::microseconds::max()) noexcept override
-  {
-    return true;
-  }
+      std::chrono::microseconds timeout = std::chrono::microseconds::max()) noexcept override;
 
 private:
   void InitializeLocalEndpoint();
 
 private:
   // The configuration options associated with this exporter.
-  bool isShutdown_ = false;
+  bool is_shutdown_ = false;
   ZipkinExporterOptions options_;
   std::shared_ptr<opentelemetry::ext::http::client::HttpClientSync> http_client_;
   opentelemetry::ext::http::common::UrlParser url_parser_;
   nlohmann::json local_end_point_;
+
+  // For testing
+  friend class ZipkinExporterTestPeer;
+  /**
+   * Create an ZipkinExporter using the specified thrift sender.
+   * Only tests can call this constructor directly.
+   * @param http_client the http client to be used for exporting
+   */
+  ZipkinExporter(std::shared_ptr<opentelemetry::ext::http::client::HttpClientSync> http_client);
+
+  mutable opentelemetry::common::SpinLockMutex lock_;
+  bool isShutdown() const noexcept;
 };
 }  // namespace zipkin
 }  // namespace exporter